requirements.txt
openai==1.30.1
opencv-python
requests
pillow
sewar
screeninfo
pyperclip
scikit-image
gradio
langdetect
fasttext-langdetect
profanity-check
alt-profanity-check
scikit-learn
joblib
python-dotenv